# swagger.yaml
openapi: 3.0.0
info:
  title: Stock Trading API
  description: API for fetching stock data, analyzing it with AWS Sage, and storing/retrieving results.
  version: 1.0.0
paths:
  /v1/retrieve/{username}/{stockname}:
    get:
      summary: Retrieves data for a stock
      description: Retrieves data for a stock from S3 bucket
      parameters:
        - name: stockname
          in: path
          description: The stock's name
          required: true
          schema:
            type: string
        - name: username
          in: path
          description: the user's username
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Data successfully retrieved from s3
          content:
            application/json:
              schema:
                type: object
                properties:
                  data_source:
                    type: string
                    example: "yahoo_finance"
                  dataset_type:
                    type: string
                    example: "Daily stock data"
                  dataset_id:
                    type: string
                  time_object:
                    type: object
                    properties:
                      timestamp:
                        type: string
                        format: date-time
                        example: "2025-03-21T12:34:56"
                      timezone:
                        type: string
                        example: "GMT+11"
                  stock_name:
                    type: string
                  events:
                    type: string
        '400':
          description: Bad Request - Invalid parameters.
        '401':
          description: Stock data not found
        '500':
          description: Internal server error.
  /v1/list/{username}/:
    get:
      summary: Lists all user's stocks
      description: Lists all stocks user has collected data on
      parameters:
        - name: username
          in: path
          description: the user's username
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: User stocks successfully listed out
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  example: "AAPL"
        '401':
          description: User not found
        '500':
          description: Internal server error.
  /v1/delete/{username}/{filename}:
    delete:
      summary: "Delete a file from DynamoDB"
      description: "This endpoint deletes a specified file for the given username from the DynamoDB table."
      parameters:
        - name: "filename"
          in: "path"
          required: true
          schema:
            type: string
          description: Name of file to be deleted
        - name: "username"
          in: "path"
          required: true
          schema:
            type: string
          description: "The user's username"
      responses:
        '200':
          description: File deleted successfully
        '401':
          description: User not found
        '500':
          description: Internal server error.
...
components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer